internal/profile: move FromKubeconfig here to be used from other commands
